djangoProject3/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
filelock==3.12.2
idna==3.4
mysqlclient==2.2.0
Pillow==10.0.0
pipenv==2023.7.23
platformdirs==3.10.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==6.0.0
typing-extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.3
virtualenv-clone==0.5.7
django-filter~=23.2